exercises 4.20.-4.22.

+++ osa4/blogilista/controllers/blogs.js
@@ -1,15 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
 
 blogsRouter.get('/', async (request, response) => {  
   const blogs = await Blog
@@ -19,11 +9,10 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
+  const user = request.user
+  if (user === null) {
     return response.status(401).json({ error: 'token invalid' })
   }
-  const user = await User.findById(decodedToken.id)
 
   const newBlog = new Blog({
     title: body.title,
@@ -35,13 +24,26 @@ blogsRouter.post('/', async (request, response) => {
 
   const savedBlog = await newBlog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
+
   await user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  if (user === null) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blogId = request.params.id
+  const blog = await Blog.findById(blogId)
+
+  if (blog.user.toString() === user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'no permission' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
